Decrypt reasoning in frontend
